Fix build failure due to link order of libraries and also add CXXFLAGS
From http://www.network-theory.co.uk/docs/gccintro/gccintro_18.html:
The traditional behavior of linkers is to search for external functions
from left to right in the libraries specified on the command line. This
means that a library containing the definition of a function should
appear after any source files or object files which use it.