Use IPC_connectModuleNoListen instead of IPC_connectModule for publisher
Note that this breaks the build for older (< 3.9.0) versions of IPC
since they do not contain the IPC_connectModuleNoListen function.

With newer (>= 3.9.0) versions of IPC, I found that if the publisher of
a message is started before the subscriber, the subscriber does not
receive any messages. If the subscriber is started before the publisher
then all is fine.

I emailed Reid Simmons (contact person for IPC) regarding this and the
reply I received was:
Basically, the problem is that, prior to v3.9, there was a bug in IPC
such that a module sent messages to central even if no one had
subscribed to them. When a module publishes a message for the first
time, it gets information from central about the message and if there
are any handlers, and also sets up a callback for central to inform it
when new handlers subscribe. If you don't periodically do an IPC_listen,
the callbacks are not invoked. While a module could ask whether there is
a handler registered every time it sends out a message, this is
inefficient. The IPC_connectModuleNoListen basically tells the module to
send out messages every time, even if it does not know that anyone has
subscribed, just in case a new subscriber has come along since the
message was first published.